style: Puts FakeHotels into a separate file
